+++ Classes/Hooks/ThemesDomainRepositoryThemeRepositoryInitHook.php
@@ -29,71 +29,47 @@ namespace KayStrobach\Themes\Hooks;
 
 use KayStrobach\Themes\Domain\Model\Theme;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
 
 /**
  * Hooks into the theme repo to load the list of themes.
  */
 class ThemesDomainRepositoryThemeRepositoryInitHook
 {
-    /**
-     * @var array
-     *
-     * @todo find a more flexible solution
-     */
-    protected $ignoredExtensions = [
-        'themes',
-        'skinselector_content',
-        'skinpreview',
-        'templavoila',
-        'piwik',
-        'piwikintegration',
-        'templavoila_framework',
-        'be_acl',
-        'sitemgr',
-        'sitemgr_template',
-        'sitemgr_fesettings',
-        'sitemgr_fe_notfound',
-        'cal',
-        'extension_builder',
-        'coreupdate',
-        'contextswitcher',
-        'extdeveval',
-        'powermail',
-        'kickstarter',
-        'tt_news',
-        'dyncss',
-        'dyncss_less',
-        'dyncss_scss',
-        'dyncss_turbine',
-        'static_info_tables',
-        'realurl',
-    ];
 
     /**
-     * hook function.
+     * Add all available themes to the Themes repository
      *
-     * @param $params
-     * @param $pObj
-     *
-     * @return void
-     *
-     * @todo add a more explaining description why this hook is required
+     * @param array $params
+     * @param $pObj \KayStrobach\Themes\Domain\Repository\ThemeRepository
+     * @throws \Exception
      */
     public function init(&$params, $pObj)
     {
-        // exclude extensions, which are not worth to check them
+        //
+        // Themes configuration
+        $extensionConfiguration = GeneralUtility::makeInstance(ExtensionConfiguration::class);
+        $themesConfiguration = $extensionConfiguration->get('themes');
+        $startWith = $themesConfiguration['themeExtensionsStartWith'];
+        //
+        // Get all available extensions, excluding system extensions
         $extensionsToCheck = array_diff(
             ExtensionManagementUtility::getLoadedExtensionListArray(),
-            $this->ignoredExtensions,
-            scandir(PATH_typo3.'sysext')
+            scandir(PATH_typo3 . 'sysext')
         );
-
-        // check extensions, which are worth to check
+        //
+        // Check extensions, which are worth to check
         foreach ($extensionsToCheck as $extensionName) {
-            $extPath = ExtensionManagementUtility::extPath($extensionName);
-            if (file_exists($extPath . 'Meta/theme.yaml') && (file_exists($extPath . 'Configuration/TypoScript/setup.txt') || file_exists($extPath . 'Configuration/TypoScript/setup.typoscript'))) {
-                $pObj->add(new Theme($extensionName));
+            if(trim($startWith) === '' || substr($extensionName, 0, strlen($startWith)) === $startWith) {
+                $extPath = ExtensionManagementUtility::extPath($extensionName);
+                if (file_exists($extPath . 'Meta/theme.yaml')) {
+                    if (file_exists($extPath . 'Configuration/TypoScript/setup.typoscript')) {
+                        $pObj->add(new Theme($extensionName));
+                    }
+                }
             }
         }
     }
+
 }